refactor: follow proper folder convention and add comming soon page for better ux
